+++ packages/talker_logger/lib/src/utils/formatter/extended_logger_formatter.dart
@@ -1,0 +1,30 @@
+import 'package:talker_logger/talker_logger.dart';
+
+/// This formatter makes messages colorful
+/// if this setting is enabled in the settings [TalkerLoggerSettings]
+class ExtendedLoggerFormatter implements LoggerFormatter {
+  const ExtendedLoggerFormatter();
+
+  @override
+  String fmt(LogDetails details, TalkerLoggerSettings settings) {
+    final underline = ConsoleUtils.getUnderline(
+      settings.maxLineWidth,
+      lineSymbol: settings.lineSymbol,
+      withCorner: true,
+    );
+    final topline = ConsoleUtils.getTopline(
+      settings.maxLineWidth,
+      lineSymbol: settings.lineSymbol,
+      withCorner: true,
+    );
+    final msg = details.message?.toString() ?? '';
+    final msgBorderedLines = msg.split('\n').map((e) => '│ $e');
+    if (!settings.enableColors) {
+      return '$topline\n${msgBorderedLines.join('\n')}\n$underline';
+    }
+    var lines = [topline, ...msgBorderedLines, underline];
+    lines = lines.map((e) => details.pen.write(e)).toList();
+    final coloredMsg = lines.join('\n');
+    return coloredMsg;
+  }
+}