+++ packages/talker_logger/test/talker_logger_test.dart
@@ -1,7 +1,7 @@
 import 'package:talker_logger/talker_logger.dart';
 import 'package:test/test.dart';
 
-class LogLevelLoggerFormater implements LoggerFormater {
+class LogLevelLoggerFormater implements LoggerFormatter {
   @override
   String fmt(LogDetails details, TalkerLoggerSettings settings) {
     return details.level.title;
@@ -42,7 +42,7 @@ void main() {
     expect(logger.settings.lineSymbol, '#');
 
     // ignore: unnecessary_type_check
-    expect(logger.formater is LoggerFormater, true);
+    expect(logger.formater is LoggerFormatter, true);
     // ignore: unnecessary_type_check
     expect(logger.formater is LogLevelLoggerFormater, true);
   });
@@ -66,7 +66,7 @@ void main() {
     expect(logger.settings.lineSymbol, '#');
 
     // ignore: unnecessary_type_check
-    expect(logger.formater is LoggerFormater, true);
+    expect(logger.formater is LoggerFormatter, true);
     // ignore: unnecessary_type_check
     expect(logger.formater is LogLevelLoggerFormater, true);
   });
@@ -89,7 +89,7 @@ void main() {
     expect(logger.settings.lineSymbol, '#');
 
     // ignore: unnecessary_type_check
-    expect(logger.formater is LoggerFormater, true);
+    expect(logger.formater is LoggerFormatter, true);
     // ignore: unnecessary_type_check
     expect(logger.formater is LogLevelLoggerFormater, true);
   });
@@ -139,12 +139,13 @@ void main() {
       final logger = TalkerLogger(
         settings: const TalkerLoggerSettings(enableColors: false),
         output: (message) => _messages.add(message),
+        formater: const ColoredLoggerFormatter(),
       );
-      final str = '----' * 10000000;
+      final str = '────' * 1000;
       logger.log(str);
       expect(
         _messages[0],
-        '${'----' * 10000000}\n--------------------------------------------------------------------------------------------------------------',
+        '${'────' * 1000}\n──────────────────────────────────────────────────────────────────────────────────────────────────────────────',
       );
     });
   });