+++ packages/talker_logger/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+- **FEAT**: Implement ExtendedLoggerFormatter
+- **FEAT**: Upgrade ColoredLoggerFormatter
+- **FIX**: Typo Formater -> Formatter
+
 ## 1.0.0
 First stable release! 🎉